requirements.txt
admin==0.0.1
alembic==1.4.3
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
bcrypt==3.2.0
Brlapi==0.6.6
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
Command==0.1.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
dnspython==2.0.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.1.3
MarkupSafe==1.1.1
netifaces==0.10.4
numpy==1.19.1
oauth==1.0.1
olefile==0.45.1
pbr==5.5.0
pexpect==4.2.1
Pillow==5.1.0
pipenv==2020.6.2
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyTelegramBotAPI==3.7.1
python-apt==1.6.5+ubuntu0.4
python-dateutil==2.8.1
python-debian==0.1.32
python-editor==1.0.4
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.7.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
SQLAlchemy==1.3.20
system-service==0.3
systemd-python==234
testresources==2.0.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
virtualenv==20.0.21
virtualenv-clone==0.5.4
wadllib==1.3.2
Werkzeug==1.0.1
WTForms==2.3.3
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
